--- .gitignore
@@ -1,14 +1,0 @@
-.DS_Store
-.DS_Store?
-._*
-.Spotlight-V100
-.Trashes
-ehthumbs.db
-Thumbs.db
-
-node_modules/
-deploy/
-json_relay_node/node_modules/
-dist/node_modules/
-_ens/
-*.zip
