+++ app/scripts/globalFuncs.js
@@ -226,19 +226,15 @@ globalFuncs.removeTokenFromLocal = function(symbol, tokenObj) {
             break;
         }
 };
-globalFuncs.localStorage = {
-        /* Not very elegant … pls hlp */
-        cookieStorage: false,
 
+
+globalFuncs.localStorage = {
         isAvailable: function() {
             // return typeof localStorage != "undefined";
+            // return globalFuncs.storageAvailable('localStorage');
             
-            // console.log("isAvailable() globalFuncs.storageAvailable: " + globalFuncs.storageAvailable('localStorage'));
-            if(!globalFuncs.storageAvailable('localStorage') && this.cookieStorage == false) {
-                this.cookieStorage = true;
-                globalFuncs.createCookieLocalStorage();
-            }
-            return globalFuncs.storageAvailable('localStorage');
+            // Polyfilled if not available/accessible 
+            return true;
         },
         setItem: function(key, value) {
             if (this.isAvailable()) {
@@ -258,6 +254,7 @@ globalFuncs.localStorage = {
 
 
 /* Check for 'localStorage' or 'sessionStorage' */
+/*
 globalFuncs.storageAvailable = function(type) {
     try {
         var storage = window[type],
@@ -280,73 +277,9 @@ globalFuncs.storageAvailable = function(type) {
             // acknowledge QuotaExceededError only if there's something already stored
             storage.length !== 0;
     }
-}
-
-/* Create fallback for localStorage based on cookies */
-globalFuncs.createCookieLocalStorage = function() {
-    // console.info("creating localStorage fallback");
-    if (!globalFuncs.localStorage.isAvailable()) {
-      Object.defineProperty(window, "localStorage", new (function () {
-        var aKeys = [], oStorage = {};
-        Object.defineProperty(oStorage, "getItem", {
-          value: function (sKey) { return sKey ? this[sKey] : null; },
-          writable: false,
-          configurable: false,
-          enumerable: false
-        });
-        Object.defineProperty(oStorage, "key", {
-          value: function (nKeyId) { return aKeys[nKeyId]; },
-          writable: false,
-          configurable: false,
-          enumerable: false
-        });
-        Object.defineProperty(oStorage, "setItem", {
-          value: function (sKey, sValue) {
-            if(!sKey) { return; }
-            document.cookie = escape(sKey) + "=" + escape(sValue) + "; expires=Tue, 19 Jan 2038 03:14:07 GMT; path=/";
-          },
-          writable: false,
-          configurable: false,
-          enumerable: false
-        });
-        Object.defineProperty(oStorage, "length", {
-          get: function () { return aKeys.length; },
-          configurable: false,
-          enumerable: false
-        });
-        Object.defineProperty(oStorage, "removeItem", {
-          value: function (sKey) {
-            if(!sKey) { return; }
-            document.cookie = escape(sKey) + "=; expires=Thu, 01 Jan 1970 00:00:00 GMT; path=/";
-          },
-          writable: false,
-          configurable: false,
-          enumerable: false
-        });
-        this.get = function () {
-          var iThisIndx;
-          for (var sKey in oStorage) {
-            iThisIndx = aKeys.indexOf(sKey);
-            if (iThisIndx === -1) { oStorage.setItem(sKey, oStorage[sKey]); }
-            else { aKeys.splice(iThisIndx, 1); }
-            delete oStorage[sKey];
-          }
-          for (aKeys; aKeys.length > 0; aKeys.splice(0, 1)) { oStorage.removeItem(aKeys[0]); }
-          for (var aCouple, iKey, nIdx = 0, aCouples = document.cookie.split(/\s*;\s*/); nIdx < aCouples.length; nIdx++) {
-            aCouple = aCouples[nIdx].split(/\s*=\s*/);
-            if (aCouple.length > 1) {
-              oStorage[iKey = unescape(aCouple[0])] = unescape(aCouple[1]);
-              aKeys.push(iKey);
-            }
-          }
-          return oStorage;
-        };
-        this.configurable = false;
-        this.enumerable = true;
-      })());
-    }
     
 }
+*/
 
     // globalFuncs.getUrlParameter = function getUrlParameter(url) {
     //   // get query string from url (optional) or window
