+++ app/scripts/main.js
@@ -1,4 +1,5 @@
 'use strict';
+require('./localStoragePolyfill');
 var IS_CX = false;
 if (typeof chrome != 'undefined') IS_CX = chrome.windows === undefined ? false : true;
 var angular                  = require('angular');
